+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Craft CMS 4
 
+## Unreleased (4.10)
+
+- The `collect()` Twig function now returns a `craft\elements\ElementCollection` instance if all of the items are elements.
+
 ## Unreleased
 
 - Fixed a bug where the `db/backup` command could fail on Windows. ([#15090](https://github.com/craftcms/cms/issues/15090))